[Chat] Use the same component for messages in timeline and edit views
